# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PetCare AI - Pet Operating System",
    "description": "AI-powered platform for pet health tracking, vet collaboration, and emergency response",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/medical/vaccinations": {
      "post": {
        "tags": [
          "Medical"
        ],
        "summary": "Add vaccination record [Owner/Vet/Caretaker]",
        "description": "**Roles allowed:** owner, caretaker, shelter, vet, admin\n- Owners log vaccinations for their pets.\n- Vets can log for any pet.",
        "operationId": "add_vaccination_api_medical_vaccinations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaccineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/medical/vaccinations/{pet_id}": {
      "get": {
        "tags": [
          "Medical"
        ],
        "summary": "List vaccinations [Owner/Vet]",
        "description": "**Roles allowed:** all authenticated users",
        "operationId": "get_vaccinations_api_medical_vaccinations__pet_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/vaccinations/{vac_id}": {
      "delete": {
        "tags": [
          "Medical"
        ],
        "summary": "Delete vaccination [Vet/Admin only]",
        "description": "**Roles allowed:** vet, admin only",
        "operationId": "delete_vaccination_api_medical_vaccinations__vac_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vac_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vac Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/vaccinations/{pet_id}/due": {
      "get": {
        "tags": [
          "Medical"
        ],
        "summary": "Get upcoming due vaccinations [All roles]",
        "description": "Returns vaccinations due within `days_ahead` days (default 30).",
        "operationId": "due_vaccinations_api_medical_vaccinations__pet_id__due_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          },
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days Ahead"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/medications": {
      "post": {
        "tags": [
          "Medical"
        ],
        "summary": "Add medication [Owner/Vet]",
        "description": "**Roles allowed:** owner, caretaker, shelter, vet, admin\n- Owners add OTC medications.\n- Vets add prescription medications.",
        "operationId": "add_medication_api_medical_medications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/medical/medications/{pet_id}": {
      "get": {
        "tags": [
          "Medical"
        ],
        "summary": "List medications [All roles]",
        "description": "Set `active_only=false` to include inactive/completed medications.",
        "operationId": "get_medications_api_medical_medications__pet_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/medications/{med_id}": {
      "put": {
        "tags": [
          "Medical"
        ],
        "summary": "Update medication [Owner/Vet]",
        "description": "Update dosage, frequency, end date, or deactivate a medication.",
        "operationId": "update_medication_api_medical_medications__med_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "med_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Med Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/medications/{med_id}/deactivate": {
      "put": {
        "tags": [
          "Medical"
        ],
        "summary": "Stop medication [Owner/Vet]",
        "description": "Mark a medication as completed/stopped.",
        "operationId": "deactivate_medication_api_medical_medications__med_id__deactivate_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "med_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Med Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/medical/appointments": {
      "post": {
        "tags": [
          "Medical"
        ],
        "summary": "Book appointment [Owner/Caretaker/Shelter]",
        "description": "**Roles allowed:** owner, caretaker, shelter, vet, admin\n- Owners/caretakers book appointments for their pets.\n- Vets can also create appointments on behalf of owners.",
        "operationId": "create_appointment_api_medical_appointments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/medical/appointments/{pet_id}": {
      "get": {
        "tags": [
          "Medical"
        ],
        "summary": "List appointments [All roles]",
        "description": "Filter by status with `?status=pending|confirmed|completed|cancelled`",
        "operationId": "get_appointments_api_medical_appointments__pet_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VaccineCreate": {
        "properties": {
          "pet_id": {
            "type": "integer",
            "title": "Pet Id"
          },
          "vaccine_name": {
            "type": "string",
            "title": "Vaccine Name"
          },
          "administered_date": {
            "type": "string",
            "format": "date-time",
            "title": "Administered Date"
          },
          "next_due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Due Date"
          },
          "administered_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Administered By"
          },
          "batch_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Number"
          }
        },
        "type": "object",
        "required": [
          "pet_id",
          "vaccine_name",
          "administered_date"
        ],
        "title": "VaccineCreate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}